Moved service classes together
