+++ src/main/java/com/waracle/cakemgr/service/SeedDataService.java
@@ -1,17 +1,13 @@
-package com.waracle.cakemgr.apiserver.service;
+package com.waracle.cakemgr.service;
 
-import com.waracle.cakemgr.apiserver.utilities.SeedDataRetriever;
 import com.waracle.cakemgr.entities.SeedDataCake;
-import com.waracle.cakemgr.persistence.CakeEntity;
-import com.waracle.cakemgr.persistence.CakeRepository;
+import com.waracle.cakemgr.persistence.*;
 import org.springframework.beans.factory.annotation.Autowired;
-import org.springframework.context.annotation.Profile;
-import org.springframework.stereotype.Component;
+import org.springframework.stereotype.Service;
 
 import javax.annotation.PostConstruct;
 
-@Component
-@Profile("apiserver")
+@Service
 public class SeedDataService {
 
     @Autowired
